backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
python-dotenv
tensorflow
numpy
pillow